@@ -1,14 +1,20 @@
+import { CadastrarReceitasComponent } from './components/receitas/cadastrar-receitas/cadastrar-receitas.component';
+import { ListarReceitasComponent } from './components/receitas/listar-receitas/listar-receitas.component';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { HomeComponent } from './components/home/home.component';
+import { CadastrarProdutoComponent } from './components/produtos/cadastrar-produto/cadastrar-produto.component';
 import { ListarProdutosComponent } from './components/produtos/listar-produtos/listar-produtos.component';
-import { CadastrarProdutosComponent } from './components/produtos/cadastrar-produtos/cadastrar-produtos.component';
+import { HomeComponent } from './components/home/home.component';
+import { AtualizarProdutoComponent } from './components/produtos/atualizar-produto/atualizar-produto.component';
 
 const routes: Routes = [
-  { path: '', component:HomeComponent },
-  { path: 'produtos', component:ListarProdutosComponent},
-  { path: 'produtos/cadastrar', component: CadastrarProdutosComponent}
+  {path:'', component:HomeComponent},
+  {path:'produtos', component:ListarProdutosComponent},
+  {path:'produtos/cadastrar', component:CadastrarProdutoComponent},
+  {path:'produtos/atualizar/:id', component:AtualizarProdutoComponent},
+  {path:'receitas', component:ListarReceitasComponent},
+  {path:'receitas/cadastrar', component:CadastrarReceitasComponent}
 ];
 
 @NgModule({
