@@ -1,27 +1,25 @@
-import { IProduto } from './../../../model/IProduto.model';
 import { ProdutosService } from './../../../services/produtos.service';
+import { IProduto } from './../../../model/IProduto.model';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-listar-produtos',
   templateUrl: './listar-produtos.component.html',
-  styleUrls: ['./listar-produtos.component.css']
+  styleUrls: ['./listar-produtos.component.css'],
 })
-
 export class ListarProdutosComponent implements OnInit {
-
-
   listaProdutos: IProduto[] = [];
 
-  constructor(private produtosService: ProdutosService) {
+  constructor(private produtosService: ProdutosService) {}
 
+  ngOnInit(): void {
+    this.carregarProdutos();
   }
-  carregarProdutos(): void{
-    this.produtosService.buscarTodos().subscribe(retorno => {
+
+  carregarProdutos(): void {
+    this.produtosService.buscarTodos().subscribe(retorno =>{
       this.listaProdutos = retorno;
     })
   }
-  ngOnInit(): void{
-    this.carregarProdutos();
-  }
+
 }
