@@ -1,4 +1,4 @@
-export interface IProduto {
+export interface IProduto{
   id?: number;
   nome: string;
   validade: Date;
